+++ src/entities/sparks.js
